🍖🐟 Checkpoint ./server.js:850652/181 ./data/steps.json:850652/53 ./commands/close.js:850652/1280

+++ commands/close.js
@@ -1,4 +1,13 @@
+const fs = require("fs");
+const low = require("lowdb");
+const FileSync = require("lowdb/adapters/FileSync");
+
 exports.run = (client, message, args, tools) => {
+  const stepsdb = new FileSync("./data/steps.json");
+  const db = low(stepsdb);
   if (!message.channel.name.includes("ticket-")) return;
+  var data = db.find({ id: message.author.id }).value();
+  if (!data) return;
+  db.remove({ id: message.author.id }).write();
   message.channel.delete();
-}
+};
